// src/common/enum/enum.ts
export enum StatusCode {
  SUCCESS = 200,
  NO_CONTENT = 200,
  RESOURCE_CREATED = 201,
  BAD_REQUEST = 400,
  UNAUTHENTICATED = 401,
  UNAUTHORIZED = 403,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  SERVER_ERROR = 500,
}

export enum ErrorCode {
  UNAUTHENTICATED = "LMS3000",
  UNAUTHORIZED = "LMS3001",
  FORBIDDEN = "LMS3003",
  BAD_REQUEST = "LMS4000",
  INVALID_QUERY = "LMS4001",
  INVALID_PARAMS = "LMS4002",
  INVALID_BODY = "LMS4003",
  FAILED_ON_AUTHENTICATION = "LMS4004",
  FOREIGN_KEY_CONSTRAINT = "LMS4020",
  UNIQUE_CONSTRAINT = "LMS4021",
  NON_EXISTENT_RESOURCE = "LMS4022",
  RESOURCE_NOT_FOUND = "LMS4040",
  INTERNAL_SERVER_ERROR = "LMS5000",
}

export enum LocalStorageKey {
  TRANSACTION = "TRANSACTION",
}